Regression-based modeling + Core ML concepts
In order to provide predictions, machine learning (ML) algorithms identify patterns in data.  A simple example of modeling a straight-line relationship between inputs and outputs is linear regression.  Features, labels, training, and generalization are among the fundamental ideas of machine learning.  Particularly in regression tasks, evaluation metrics like as MSE, MAE, and RMSE are crucial for gauging model accuracy.  Selecting the best-performing algorithm is part of the model selection process, and distance measurements aid in the comparison of data points.  Several input variables are used in multiple linear regression to increase prediction accuracy.